+++ Sources/Mimus/Matchers/CaptureArgumentMatcher.swift
@@ -1,13 +1,13 @@
 /// Records are compared values.
 /// See https://github.com/AirHelp/Mimus/blob/master/Documentation/Capturing%20Arguments.md for more details. 
-public class CaptureArgumentMatcher: MockEquatable {
+public class CaptureArgumentMatcher: Matcher {
 
     public var capturedValues: [Any?] = []
 
     public init() { }
 
-    public func equalTo(other: Any?) -> Bool {
-        capturedValues.append(other)
+    public func matches(argument: Any?) -> Bool {
+        capturedValues.append(argument)
         return true
     }
 }