+++ Sources/Mimus/Matchers/MatcherHelpers.swift
@@ -6,7 +6,7 @@ import Foundation
 
 /// Convenience extension to reduce boilerplate code when writing custom extension matchers.
 /// See https://github.com/AirHelp/Mimus/blob/master/Documentation/Using%20Your%20Own%20Types.md for details. 
-public extension MockEquatable {
+public extension Matcher {
     func compare<T: Equatable>(other: T?) -> Bool {
         guard let equatableSelf = self as? T, let other = other else {
             return false