+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-# Next
-Nothing yet!
+# Next (2.0.0)
+- **BREAKING** `MockEquatable` has been renamed to `Matcher` to more closely express functionality provided (there is a typealias for backwards compatibility _but_ it might still be breaking in some cases)
+- **BREAKING** `Mock.storage` type has been changed from `[RecordedCall]` to `Storage` to better support future features (like e.g. the recording of return values)
+- **BREAKING** Calling verify will now reset previously recorded values
+- Added option to record values that should be returned for specific invocations
+
+# 1.1.5
+- SPM Support
 
 # 1.1.4
 - Mismatched arguments are now included in failure message (thanks @Eluss)